chore: Moves event data to context
