+++ src/App.jsx
@@ -4,7 +4,6 @@ import EventMain from "./components/EventMain";
 import "./styles.scss";
 
 function App() {
-  console.log("App function");
   return (
     <div className="app">
       <EventMain />
